Split to swiftide-indexing
