+++ swiftide-core/src/lib.rs
@@ -6,3 +6,8 @@ pub mod type_aliases;
 
 pub use traits::*;
 pub use type_aliases::*;
+
+pub mod indexing {
+    pub use crate::indexing_stream::IndexingStream;
+    pub use crate::node::*;
+}
