+++ swiftide-indexing/src/lib.rs
@@ -1,3 +1,6 @@
-fn main() {
-    println!("Hello, world!");
-}
+pub mod loaders;
+pub mod persist;
+pub mod pipeline;
+pub mod transformers;
+
+pub use pipeline::Pipeline;
