+++ swiftide-indexing/src/loaders/mod.rs
