+++ swiftide-indexing/src/transformers/metadata_keywords.rs
@@ -1,7 +1,7 @@
 //! Extract keywords from a node and add them as metadata
 use std::sync::Arc;
 
-use crate::{indexing::Node, prompt::PromptTemplate, SimplePrompt, Transformer};
+use swiftide_core::{indexing::Node, prompt::PromptTemplate, SimplePrompt, Transformer};
 use anyhow::Result;
 use async_trait::async_trait;
 use derive_builder::Builder;
@@ -106,7 +106,7 @@ impl Transformer for MetadataKeywords {
 
 #[cfg(test)]
 mod test {
-    use crate::MockSimplePrompt;
+    use swiftide_core::MockSimplePrompt;
 
     use super::*;
