+++ swiftide-indexing/src/transformers/prompts/metadata_summary.prompt.md
