+++ swiftide/src/lib.rs
@@ -56,11 +56,6 @@
 //! Either use the 'all' feature flag (not recommended), or enable the integrations that you need.
 //! Each integration has a similarly named feature flag.
 
-pub mod indexing;
-pub mod loaders;
-pub mod persist;
-pub mod transformers;
-
 #[doc(inline)]
 pub use swiftide_core::prompt;
 #[doc(inline)]
@@ -76,15 +71,7 @@ pub mod integrations {
     pub use swiftide_integrations::*;
 }
 
-/// Deprecated re-export of `indexing`, use that instead.
-#[deprecated(
-    since = "0.6.0",
-    note = "Renamed references of Indexing to Indexing for more appropriate naming. Will be removed in a future release."
-)]
-pub mod ingestion {
-    pub use crate::indexing::*;
-
-    pub use crate::indexing::IndexingStream;
-    pub use crate::indexing::Node;
-    pub use crate::indexing::Pipeline;
+pub mod indexing {
+    pub use swiftide_core::indexing::*;
+    pub use swiftide_indexing::*;
 }
